refactor broker changes for DCF flow
